requirements.txt
amqp==1.4.6
anyjson==0.3.3
Babel==1.3
billiard==3.3.0.20
celery==3.1.18
cffi==0.9.2
cliff==1.12.0
cmd2==0.6.8
cryptography==0.9
Django==1.7.8
django-appconf==1.0.1
django-celery==3.1.16
django-compressor==1.5
django-openstack-auth==1.3.1
djangorestframework==3.1.1
djproxy==2.1.0
python-ldap==2.4.20
django-auth-ldap==1.2.6
ecdsa==0.13
enum34==1.0.4
Fabric==1.10.2
idna==1.1
ipaddress==1.0.7
iso8601==0.1.10
jsonpatch==1.11
jsonpointer==1.9
jsonschema==2.4.0
kombu==3.0.26
msgpack-python==0.4.6
MySQL-python==1.2.5
netaddr==0.7.14
netifaces==0.10.4
oslo.config==1.11.0
oslo.i18n==1.6.0
oslo.serialization==1.5.0
oslo.utils==1.5.0
paramiko==1.15.2
pbr==0.11.0
prettytable==0.7.2
pyasn1==0.1.7
pycparser==2.13
pycrypto==2.6.1
pyOpenSSL==0.15.1
pyparsing==2.0.3
python-ceilometerclient==1.2.0
python-cinderclient==1.2.1
python-glanceclient==0.18.0
python-heatclient>=0.3.0,<0.5.0
python-swiftclient>=2.2.0,<2.5.0
python-keystoneclient==1.4.0
python-neutronclient==2.5.0
python-novaclient==2.25.0
pytz==2015.4
requests==2.7.0
simplejson==3.6.5
six==1.9.0
stevedore==1.4.0
warlock==1.1.0
Pillow==2.8.2
django-simple-captcha==0.4.6
